This resolved #3, resolved #4, resolved #5, resolved #6, resolved #8, resolved #10
