--- src/Main.java
@@ -1,8 +1,0 @@
-public class Main {
-    //Made others change
-    public static void main(String[] args) {
-        EmployeeRepository employeeRepository = new EmployeeRepository();
-        EmployeeService employeeService = new EmployeeService(employeeRepository);
-        employeeService.run();
-    }
-}
