--- src/exeptions/FileNotFoundCustomException.java
@@ -1,7 +1,0 @@
-package exeptions;
-//Made others change
-public class FileNotFoundCustomException extends Exception {
-    public FileNotFoundCustomException(String message) {
-        super(message);
-    }
-}
