--- src/exeptions/InvalidEntryParametersException.java
@@ -1,7 +1,0 @@
-package exeptions;
-//Made others change
-public class InvalidEntryParametersException extends Exception {
-    public InvalidEntryParametersException(String message) {
-        super(message);
-    }
-}
