--- src/interfaces/IEmployee.java
@@ -1,7 +1,0 @@
-package interfaces;
-//Made others change
-public interface IEmployee {
-    String getName();
-    int getAge();
-    double getLengthOfService();
-}
