--- src/interfaces/IEmployeeRepository.java
@@ -1,17 +1,0 @@
-package interfaces;
-//Made others change
-import java.util.List;
-
-public interface IEmployeeRepository {
-    List<IEmployee> getEmployeeList();
-
-    void addEmployee(IEmployee employee);
-
-    double averageAgeOfEmployees();
-
-    double averageLengthOfServiceOfEmployees();
-
-    double maximumLengthOfServiceOfEmployee();
-
-    String mostCommonCharactersInEmployeesNames();
-}
