+++ src/main/java/Main.java
@@ -1,0 +1,12 @@
+package main.java;
+
+import main.java.interfaces.IEmployeeService;
+
+public class Main {
+
+    public static void main(String[] args) {
+        EmployeeRepository employeeRepository = new EmployeeRepository();
+        IEmployeeService employeeService = new EmployeeService(employeeRepository);
+        employeeService.initialize();
+    }
+}
