+++ src/main/java/exeptions/NoEmployeesException.java
@@ -1,5 +1,5 @@
-package exeptions;
-//Made others change
+package main.java.exeptions;
+
 public class NoEmployeesException extends Exception{
     public NoEmployeesException(String message) {
         super(message);
