+++ src/main/java/interfaces/IEmployeeService.java
@@ -1,0 +1,18 @@
+package main.java.interfaces;
+
+import java.util.HashMap;
+import java.util.List;
+
+public interface IEmployeeService {
+    void initialize();
+
+    double averageAgeOfEmployees();
+
+    double averageLengthOfServiceOfEmployees();
+
+    double maximumLengthOfServiceOfEmployee();
+
+    List<Character> mostCommonCharactersInEmployeesNames();
+
+    HashMap<Character, Integer> fillEmployeeNamesInToMap();
+}
